Se termino la paginacion en la en todas las tablas

+++ src/app/abonos/service/data-abonos.service.ts
@@ -32,8 +32,9 @@ query paginationpayments(
   $credit_quantity:Int,
   $credit_date:DateTime,
   $quantity_payment:String,
-  $credit:Float,
-  $user:String){
+  $credit:Int,
+  $user:String
+  ){
     paginationpayments(
       start:$start,
       limit:$limit,
